requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
psycopg2>=2.8.6,<2.9
rdkit-pypi==2020.9.5.2
django-rdkit @ git+https://github.com/rdkit/django-rdkit.git
drf-spectacular>=0.15.1,<0.16
selfies==2.1.1
requests==2.31.0
numpy==1.20
scipy==1.7.3
django-simple-history==3.4.0
uwsgi>=2.0.19<2.1